Separate metric implementation from monitoring resource rendering
