+++ pkg/controller/networkaddonsconfig/networkaddonsconfig_controller.go
@@ -10,6 +10,7 @@ import (
 	"time"
 
 	monitoringv1 "github.com/coreos/prometheus-operator/pkg/apis/monitoring/v1"
+	osconfv1 "github.com/openshift/api/config/v1"
 	osv1 "github.com/openshift/api/operator/v1"
 	osnetnames "github.com/openshift/cluster-network-operator/pkg/names"
 	"github.com/pkg/errors"
@@ -39,11 +40,10 @@ import (
 	"github.com/kubevirt/cluster-network-addons-operator/pkg/apply"
 	"github.com/kubevirt/cluster-network-addons-operator/pkg/controller/statusmanager"
 	"github.com/kubevirt/cluster-network-addons-operator/pkg/eventemitter"
-	"github.com/kubevirt/cluster-network-addons-operator/pkg/monitoring"
+	"github.com/kubevirt/cluster-network-addons-operator/pkg/monitoring/metrics"
 	"github.com/kubevirt/cluster-network-addons-operator/pkg/names"
 	"github.com/kubevirt/cluster-network-addons-operator/pkg/network"
 	"github.com/kubevirt/cluster-network-addons-operator/pkg/util/k8s"
-	osconfv1 "github.com/openshift/api/config/v1"
 )
 
 // ManifestPath is the path to the manifest templates
@@ -221,7 +221,7 @@ func (r *ReconcileNetworkAddonsConfig) Reconcile(ctx context.Context, request re
 			r.stopTrackingObjects()
 
 			if r.clusterInfo.MonitoringAvailable {
-				monitoring.ResetMonitoredComponents()
+				metrics.ResetMonitoredComponents()
 			}
 
 			// Owned objects are automatically garbage collected. Return and don't requeue
@@ -313,7 +313,7 @@ func (r *ReconcileNetworkAddonsConfig) Reconcile(ctx context.Context, request re
 	r.statusManager.SetFromPods()
 
 	if r.clusterInfo.MonitoringAvailable {
-		monitoring.TrackMonitoredComponents(&networkAddonsConfig.Spec, r.statusManager)
+		metrics.TrackMonitoredComponents(&networkAddonsConfig.Spec, r.statusManager)
 	}
 
 	// Kubernetes sometimes fails to apply objects while we remove and recreate
